Warnings fix :warning:

+++ Shurikenjutsu/NinjaServer/PlayerManager.cpp
@@ -11,7 +11,7 @@ PlayerManager::~PlayerManager()
 
 bool PlayerManager::Initialize(RakNet::RakPeerInterface *p_serverPeer, std::string p_levelName)
 {
-	m_playerHealth = CHARACTAR_KATANA_SHURIKEN_HEALTH;
+	m_playerHealth = (int)CHARACTAR_KATANA_SHURIKEN_HEALTH;
 	m_gcd = ALL_AROUND_GOLOBAL_COOLDOWN;
 	m_serverPeer = p_serverPeer;
